requirements.txt
Django==3.2.16
django-cors-headers==3.13.0
djangorestframework==3.13.1
gunicorn==20.1.0
pytz==2022.1
psycopg2-binary==2.9.3
dj-database-url==1.0.0
python-decouple==3.6
whitenoise==6.2.0


